requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
multidict==4.7.6
Pillow==7.2.0
pip-tools==5.3.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
typing-extensions==3.7.4.3
yarl==1.5.1
